+++ constants.py
@@ -1,0 +1,17 @@
+COLOR_CYCLE = [
+    "#0B74FF",
+    "#12C37E",
+    "#FF7F0E",
+    "#D62728",
+    "#9467BD",
+    "#8C564B",
+    "#E377C2",
+    "#7F7F7F",
+]
+
+CONFIG_FILE = "config.yaml"
+CONTROL_FMT = "<4f"
+UPDATE_MS = 10
+N_WINDOW_SEC = 10
+SAMPLE_RATE_HZ = 100
+
